Remove start page parts of content item presenter
* Refactor LicenceFinderFormContentItemPresenter into
LicenceFinderContentItemPresenter
