--- app/presenters/licence_finder_form_content_item_presenter.rb
@@ -1,12 +1,0 @@
-class LicenceFinderFormContentItemPresenter < LicenceFinderContentItemPresenter
-  attr_reader :base_path, :content_id
-
-  def initialize(base_path, content_id)
-    @base_path = base_path
-    @content_id = content_id
-  end
-
-  def route_type
-    'prefix'
-  end
-end
