+++ config/initializers/application_metadata.rb
@@ -1,6 +1,5 @@
 APPLICATION_METADATA = {
   slug: APP_SLUG,
-  content_id: "69af22e0-da49-4810-9ee4-22b4666ac627",
   title: "Licence Finder",
   description: "Find out which licences you might need for your activity or business."
 }
