+++ spec/presenters/licence_finder_content_item_presenter_spec.rb
@@ -1,11 +1,11 @@
 require 'spec_helper'
 
 RSpec.describe LicenceFinderContentItemPresenter do
-  let(:subject) { LicenceFinderContentItemPresenter.new }
+  let(:subject) { LicenceFinderContentItemPresenter.new("/licence-finder/sectors", "4ade13fa-7e79-4bee-b809-61dbe5c3aa22") }
 
   describe "#base_path" do
     it "has the correct base path" do
-      expect(subject.base_path).to eq "/licence-finder"
+      expect(subject.base_path).to eq "/licence-finder/sectors"
     end
   end
 
@@ -18,14 +18,14 @@ RSpec.describe LicenceFinderContentItemPresenter do
       expect(subject.payload[:title]).to eq "Licence Finder"
     end
 
-    it 'uses an exact route' do
-      expect(subject.payload[:routes].first[:type]).to eql('exact')
+    it 'uses a prefix route' do
+      expect(subject.payload[:routes].first[:type]).to eql('prefix')
     end
   end
 
   describe '#content_id' do
     it 'has the expected content_id' do
-      expect(subject.content_id).to eql("69af22e0-da49-4810-9ee4-22b4666ac627")
+      expect(subject.content_id).to eql("4ade13fa-7e79-4bee-b809-61dbe5c3aa22")
     end
   end
 end
